requirements.txt
streamlit
langchain
langchain-community
langchain-groq
python-dotenv
faiss-cpu